Filling implemented in Circle,You can also transalate and Rotate the filled Objects

+++ bresenham.h
@@ -9,10 +9,10 @@ using namespace std;
 class Bresenham : public Object
 {
 public:
-    Bresenham(float* color,int thickness,string pattern);
+    Bresenham(GLubyte* color,int thickness,string pattern);
     void BslopeGT1(int steps, int x, int y, int dx, int dy,int startX,int endX,int startY,int endY);
     void BslopeLT1(int steps, int x, int y, int dx, int dy,int startX,int endX,int startY,int endY);
     void draw(int startX,int endX,int startY,int endY,int width,int height);
-    void reDrawSelectedObject(float*,int);
+    void reDrawSelectedObject(GLubyte*,int);
 };
 #endif // BRESENHAM
