+++ color.h
@@ -7,6 +7,6 @@ using namespace std;
 class Color
 {
 public:
-    static float RED[3],GREEN[3],BLUE[3],GRAY[3],YELLOW[3],NAVYBLUE[3],BLACK[3];
+    static GLubyte RED[3],GREEN[3],BLUE[3],GRAY[3],YELLOW[3],NAVYBLUE[3],BLACK[3];
 };
 #endif // COLOR
