+++ liang.cpp
@@ -2,7 +2,7 @@
 
 
 extern list<Object*> allObjects;
-extern float *CURRENTCOLOR;
+extern GLubyte *CURRENTCOLOR;
 extern string CURRENTPATTERN;
 extern int CURRENTTHICKNESS;
