+++ midpoint.h
@@ -9,11 +9,11 @@ class MidPoint : public Object
 {
 
 public:
-    MidPoint(float* color,int thickness,string pattern);
+    MidPoint(GLubyte* color,int thickness,string pattern);
     void MslopeGT1(int steps, int x, int y, int dx, int dy,int startX,int endX,int startY,int endY);
     void MslopeLT1(int steps, int x, int y, int dx, int dy,int startX,int endX,int startY,int endY);
     void draw(int startX,int endX,int startY,int endY,int width,int height);
-    void reDrawSelectedObject(float*,int);
+    void reDrawSelectedObject(GLubyte*,int);
 };
 
 #endif // MIDPOINT
