+++ symmetricDDA.h
@@ -9,9 +9,9 @@ using namespace std;
 class SymmetricDDA : public Object
 {
 public:
-    SymmetricDDA(float* color,int thickness,string pattern);
+    SymmetricDDA(GLubyte* color,int thickness,string pattern);
     void draw(int startX,int endX,int startY,int endY,int width,int height);
-    void reDrawSelectedObject(float*,int);
+    void reDrawSelectedObject(GLubyte*,int);
 };
